APIを分解
